Change int to float and added unit tests

+++ src/Service/Calculator/CalculatorInterface.php
@@ -4,5 +4,5 @@ namespace App\Service\Calculator;
 
 interface CalculatorInterface
 {
-    public function calculate(int $first, int $second): int;
+    public function calculate(float $first, float $second): float;
 }
